audio_rag/requirements.txt
google-genai>=1.7.0
google-api-core>=2.19.0
chromadb>=0.6.3
python-dotenv>=1.0.0
tqdm>=4.66.0
